Enhance internal logged messages to report referer and user-agent.
This is a backport of 6e3fd97d832671678ed48df86036f2979b11f515,
https://github.com/portante/swift/commit/6e3fd97d832671678ed48df86036f2979b11f515.

Pass the referering URL between internal servers, and set the user-agent to be
the server type (obj-server, container-server) with the process PID. In
conjunction with the transaction ID, it helps to track down which PID from a
given system was responsible for initiating the request and what that server
was working on to make this request.

This has been helpful in tracking down interactions between object, container
and account servers once the proxy server has forwarded the request.

Signed-off-by: Peter Portante <peter.portante@redhat.com>