Ween the object auditor off of the DiskFile class
Given we are going in the direction that the DiskFile class is the
abstraction layer for the needs of the REST API layer, it seems that
the auditor should not use the same methods in order to truly
"audit". Instead, only the low-level on-disk file logic for the data
directory is shared, and all other verifications are performed using
direct inspections of the data.

As a result, the auditor performs a few more double checks to make
sure the "name" metadata attribute results in the same object data
file; adds some explicit error checking for quarantinable errors (bad
metadata, name does not hash back to the same data directory, unable
to open the file, unable to fstat the file). However, while the
current code will read the entire file only to find out that it does
not have a stored ETag, this code will quarantine an object that is
missing an ETag.

We also use this opportunity to bring the test coverage to 99%.

Possible related work:

  * Check that the X-Timestamp metadata value matches the given name
    of the object data file

Change-Id: I224e42d06ac49495a59e965502bcf3f323db1929
Signed-off-by: Peter Portante <peter.portante@redhat.com>