Drop knowledge of db from node servers
Since the draft of this patch was posted, a promising approach
was suggested by ondisk.py. It allows us to not to attempt to
shoehorn the use of replicators and auditors into the broker API
which knows of no databases. Accordingly, this version is more
aggressive and not only gets rid of ".db" suffix, but the
whole hash and path calculations in servers.

Change-Id: I79a8c7bb82a20554080be06b211a8e06edc51c6d