Remove thorny os.path.exists(broker.db_file)
This was known as "clean up the initialise method" (in order to
get rid of db_file), but unfortunately this requires redoing
the deleting of account and container entities as well, because
just one os.path.exists() was in an unlucky place. This opens
a small can of worms.

Doubly unfortunately, however, is that this patch badly conflicts
with the loader ("get_backend()"). This is because we are adding
exceptions and there's no way to load them alongside the back-end
broker class. There isn't one, right? So, we're likely to change
this bunch of new exceptions to the hated returned error codes.

Exceptions are a small detail, however. The meat of the patch
remains and is quite hazardous in respect to regressions. This
needs a very careful review.

Change-Id: I54d54eb8984d6bca4be912e7451f82e11b2db6ca