Ensure a/c/o properties are always not None
We ensure the account, container and obj properities of a DiskFile
object are always not None by enforcing their use after that object
has been opened.

We also very the "name" metadata of a tombstone is valid, quarantining
it before its use.

Finally, we add documentation for the officially supported properties of a
DiskFile object.

Change-Id: I03d9fd4d987c0cd5e427288c269239ec669f7238